Moved bot commands to separate file
